refactor: 칸반Todo 컴포넌트 분리 [#61]
칸반 Todo 의 상태는 칸반에서 가질 필요가 없으며, 이를 분리하는 것이 불필요한 렌더링을 방지
할 것이라고 생각함
